requirements.txt
django-filter==25.1
django-js-asset==3.1.2
django-mptt==0.17.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
factory_boy==3.3.3
Faker==37.1.0
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.2
pluggy==1.5.0
PyJWT==2.9.0
pytest==8.3.5
pytest-django==4.11.1
pytest-factoryboy==2.7.0
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
Pillow==10.2.0
psycopg2-binary==2.9.9 